requirements.txt
Flask
openai
faiss-cpu
numpy<2.0
pandas
streamlit
requests